src/anomaly_lisa/ia_patchcore_requirement.txt
####################
## AgentIA Patchcore
####################
# Fichier requirement spécifique pour l'agent

#python==*
anomalib[full]
torchvision>=0.17.2